1) добавил работу с запросами обратной связи 2) рефакторинг кода

+++ src/main/java/pro/sky/telegrambot/listener/TelegramBotUpdatesListener.java
@@ -36,7 +36,7 @@ public class TelegramBotUpdatesListener implements UpdatesListener {
             logger.info("Processing update: {}", update);
             Message message = update.message();
             CallbackQuery callbackQuery = update.callbackQuery();
-            byte processMassageCode = 0;
+            int processMassageCode = 0;
             String logMessage = "";
             if (message != null && !message.from().isBot()) {
                 processMassageCode = telegramBotService.processMessage(message);
